# [C#] Make DirectBuffer.CheckLimit inlinable