@billsegall 2021-03-31
Looks good. That's a substantial improvement.

A separate PR for BenchmarkDotNet sounds great.

Thanks!